Add MCAP reader (#148)
* Add MCAP reader

* Emit one row per MCAP file

* Document MCAP robotics conversion plan

* Emit robotics-ready MCAP rows

* Remove MCAP raw message escape hatch

* Simplify MCAP reader internals

* Reuse reader path selection for MCAP

* Tighten MCAP sync and Zarr video defaults

* Remove redundant MCAP casts

* Harden MCAP reader edge cases

* Sort MCAP primary events for alignment

* Sort MCAP marker events for splitting

* Avoid MCAP summary probing on nonseekable streams

* Expand MCAP reader docs

* Clarify MCAP sparse synchronization docs

* Document MCAP synchronization modes

* Rename MCAP sparse mode to unsynchronized

* Fill out MCAP reader reference docs

* Move MCAP episode splitting docs

* Simplify MCAP reader row metadata

* Narrow MCAP reader API surface

* Honor explicit empty MCAP field selection

* Clarify MCAP synchronization modes

* Add MCAP sync methods

* Address MCAP review edge cases

* Reject reserved MCAP frame fields

* Clean up MCAP alignment path

* Reject missing MCAP aligned video frames

* Address MCAP row collision and empty primary videos

* Preserve duplicate primary video frames

* Refine MCAP reader episodes

* Document MCAP dataset conversion example

* Decode H264 MCAP video packets

* Add MCAP conversion example

* Address MCAP review comments

* Handle rounded MCAP video FPS

* Support fractional video FPS

* Handle 16-bit ROS MCAP images

* Extract robotics synchronization helpers

* Filter MCAP sync primary subfields

* Speed up explicit MCAP subfield decoding

* Align same-topic MCAP subfields

* Handle mixed-shape MCAP source selections

* Validate explicit MCAP fps

* Cache MCAP decoders per channel

* Stream ordered MCAP episodes without seeking

* Preserve nullable MCAP field events

* Allow missing MCAP marker topic

+++ src/refiner/video/writer.py
@@ -31,7 +31,7 @@ class WrittenVideoSegment:
     output_rel: str
     from_timestamp: float
     to_timestamp: float
-    fps: int
+    fps: float
     width: int
     height: int
     codec: str
@@ -152,7 +152,7 @@ class VideoStreamWriter:
             output_rel=self._current_output_rel,
             from_timestamp=from_timestamp,
             to_timestamp=to_timestamp,
-            fps=int(writer.fps),
+            fps=float(writer.fps),
             width=int(writer.stream.width),
             height=int(writer.stream.height),
             codec=self.transcode_config.codec,
@@ -184,7 +184,7 @@ class VideoStreamWriter:
                 output_rel=self._current_output_rel,
                 from_timestamp=from_timestamp,
                 to_timestamp=to_timestamp,
-                fps=int(probe.fps),
+                fps=float(probe.fps),
                 width=int(probe.width),
                 height=int(probe.height),
                 codec=str(probe.codec or self.transcode_config.codec),
@@ -195,9 +195,9 @@ class VideoStreamWriter:
             return WrittenVideo(segment=segment, mode="remux")
 
         fps = (
-            int(prepared.probe.fps)
+            float(prepared.probe.fps)
             if prepared.probe is not None and prepared.probe.fps is not None
-            else int(self._writer.fps)
+            else float(self._writer.fps)
             if isinstance(self._writer, TranscodeWriter)
             else None
         )
@@ -217,7 +217,7 @@ class VideoStreamWriter:
             output_rel=self._current_output_rel,
             from_timestamp=from_timestamp,
             to_timestamp=to_timestamp,
-            fps=int(writer.fps),
+            fps=float(writer.fps),
             width=int(writer.stream.width),
             height=int(writer.stream.height),
             codec=self.transcode_config.codec,
@@ -227,7 +227,7 @@ class VideoStreamWriter:
             self._rotate_writer()
         return WrittenVideo(segment=segment, mode="transcode")
 
-    def _ensure_transcode_writer(self, fps: int) -> TranscodeWriter:
+    def _ensure_transcode_writer(self, fps: float) -> TranscodeWriter:
         writer = self._writer
         if isinstance(writer, TranscodeWriter) and writer.fps == fps:
             if writer.size_bytes >= self.video_bytes_limit:
